% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylab5.R
\name{mybin}
\alias{mybin}
\title{Binomial Simulation}
\usage{
mybin(iter, n, p)
}
\arguments{
\item{iter}{Number of iterations}

\item{n}{Number of trials}

\item{p}{Probability of success on single Bernoulli Trial}
}
\value{

}
\description{
Binomial Simulation
}
\examples{
\dontrun{mybin(10,10,.5)}
}
