% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhyper.R
\name{myhyper}
\alias{myhyper}
\title{Hypergeometric Simulation}
\usage{
myhyper(iter, N, r, n)
}
\arguments{
\item{iter}{Number of iterations}

\item{N}{Number of elements}

\item{r}{Number of successes in N elements}

\item{n}{Number of drawn elements}
}
\value{

}
\description{
Hypergeometric Simulation
}
\examples{
myhyper(100,10,3,6)
}
